+++ src/Kafka/Protocol/Request/JoinGroupResponseV2.php
@@ -21,7 +21,7 @@ namespace Protocol\Kafka\Protocol\Request;
  *                                            member_id [members]
  * </pre>
  *
- * @see docs/protocol/2.8.md, sections "JoinGroup API (key 11, v0 to v3)" and "Quotas and throttle time"
+ * @see docs/protocol/2.8.md, sections "JoinGroup API (key 11, v0 to v4)" and "Quotas and throttle time"
  */
 final class JoinGroupResponseV2 extends JoinGroupResponse
 {